pyproject.toml
[tool.poetry]
name = "chroma-spec"
version = "0.1.5"
description = "chroma-spec"
authors = ["abr <abricier@users.noreply.github.com>"]
readme = "README.rst"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
colour-science = "^0.4.1"
matplotlib = "^3.5.2"
click = "^8.1.3"
imageio = "^2.20.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.0"
pre-commit = "^2.20.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
black = "^22.6.0"
coverage = "^6.4.2"
pytest-cov = "^3.0.0"
flake8-docstrings = "^1.6.0"
jsonschema = "^4.7.2"
darglint = "^1.8.1"
Sphinx = "^5.1.0"
sphinx-click = "^4.3.0"
furo = "^2022.6.21"
xdoctest = "^1.0.1"
sphinx-autobuild = "^2021.3.14"
flake8-black = "^0.3.3"
flake8-bugbear = "^22.7.1"
flake8-isort = "^4.1.2"
flake8-rst-docstrings = "^0.2.7"
pep8-naming = "^0.13.1"
myst-parser = "^0.18.0"

[tool.poetry.scripts]
chroma-spec = "chroma_spec.__main__:main"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::colour.utilities.ColourRuntimeWarning",
    ]

[tool.coverage.paths]
source = ["chroma_spec", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["chroma_spec"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
